Added argparsing and saving graph functionality
